requirements.txt
aio-pika==1.6.0
aniso8601==1.3.0
appnope==0.1.0
backports.weakref==1.0rc1
bcrypt==3.1.4
bleach==1.5.0
blinker==1.4
cachetools==2.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
configparser==3.7.1
decorator==4.1.2
Flask==0.12.2
Flask-Cors==3.0.6
Flask-DebugToolbar==0.10.1
Flask-JWT==0.3.2
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
future==0.15.2
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
joblib==0.11
Markdown==2.6.9
MarkupSafe==1.0
mongoengine==0.15.0
multidict==3.3.2
mysqlclient==1.4.1
nose==1.3.7
numpy==1.13.3
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
pika==0.11.2
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pygogo==0.10.0
pyhocon==0.3.36
PyJWT==1.4.2
pymongo==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
requests==2.18.4
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.12.0
SQLAlchemy==1.2.16
tensorflow==1.0.1
thrift==0.10.0
traitlets==4.3.2
urllib3==1.22
uvloop==0.9.1
validators==0.12.0
wcwidth==0.1.7
Werkzeug==0.12.2
yarl==0.15.0
